+++ src/main/java/com/nisovin/shopkeepers/shopobjects/entity/AbstractEntityShopObject.java
@@ -18,8 +18,22 @@ public abstract class AbstractEntityShopObject extends AbstractShopObject implem
 	public abstract AbstractEntityShopObjectType<?> getType();
 
 	@Override
-	public boolean isActive() {
+	public boolean isSpawned() {
 		return (this.getEntity() != null);
+	}
+
+	@Override
+	public boolean isActive() {
+		// Note: Entity#isValid also checks if the entity is dead.
+		// Note: The entity's valid flag may get set one tick after the chunk was unloaded. We might also want to check
+		// if the entity is still valid when handling chunk unloads, and the entity might be located in a chunk
+		// different to the one that is currently being unloaded (and that other chunk might already have been
+		// unloaded).
+		// In the past we therefore also checked if the chunk in which the entity is currently located is still loaded.
+		// However, on later versions of Spigot (late 1.14.1 and above), Entity#isValid also already checks if the chunk
+		// is currently loaded, so this is no longer required.
+		Entity entity = this.getEntity();
+		return entity != null && entity.isValid();
 	}
 
 	@Override